# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  version: "1.0.0"
  description: >
    This document defines the API for the WASAText messaging application.
    It includes endpoints for user login, updating user details,
    managing conversations, sending messages, and handling group operations.
servers:
  - url: https://localhost:3000
    description: Local development server
paths:
  /session:
    post:
      tags:
        - login
      summary: Logs in or registers a user
      description: >
        If the user does not exist, they are created and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: postSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 16
                  pattern: "^[a-zA-Z0-9_-]+$"
                  example: "alice"
      responses:
        '200':
          description: User exists; returns existing identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '201':
          description: New user created; returns newly generated identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /user/username:
    put:
      tags:
        - user
      summary: Update the current user's username
      description: Updates the username of the authenticated user.
      operationId: setMyUserName
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newName
              properties:
                newName:
                  type: string
                  minLength: 3
                  maxLength: 16
                  pattern: "^[a-zA-Z0-9_-]+$"
                  example: "newAlice"
      responses:
        '200':
          description: Username updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /user/photo:
    put:
      tags:
        - user
      summary: Update the current user's profile photo
      description: Updates the profile photo URL for the authenticated user.
      operationId: setMyPhoto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - photoUrl
              properties:
                photoUrl:
                  type: string
                  format: uri
                  pattern: "^https://.+$"
                  example: "https://example.com/photo.jpg"
      responses:
        '200':
          description: Profile photo updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token obtained from /session endpoint.
        Token should be included in the Authorization header as "Bearer <token>".
  schemas:
    UserResponse:
      type: object
      required:
        - identifier
      properties:
        identifier:
          type: string
          format: uuid
          example: "uuid-of-alice"
    SuccessResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid request parameters"
        details:
          type: array
          items:
            type: string
          example: ["Field 'name' must be between 3 and 16 characters"]
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Number of seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'